# feat: remove grouping to speed up query