Grouping was causing a big running time overhead, removing this decreases the overall query time. 

Benchmark numbers for 10k-fake.txt are 47.20 s for after the changes vs. 81.28s for before the changes. 

New
----------

Perform nilRAG...
Performance breakdown. Time to:            
 generate query embedding: 1.42 seconds             
 check clustering and (if existing) get closest centroid:                0.00 seconds            
 Number of clusters found: 0            
 encrypt query embedding: 0.00 seconds            
 ask nilDB to compute the differences: 29.91 seconds            
 transpose list: 2.01 seconds            
 decrypt: 13.24 seconds            
 get top chunks ids: 0.00 seconds            
 query top chunks: 0.47 seconds            
[
    {
        "_id": "7780ee4b-9b20-4bfd-afe2-d9be082a5e6f",
        "chunks": "Michelle Perez works at Gilmore, Vasquez and Francis as a Occupational psychologist. Michelle Perez was born on 2000-04-30 and lives at 2718 Burke Haven Suite 427, Tracieton, MD 74859."
    },
    {
        "_id": "fee6732f-c06e-4643-9f16-0a167592cfb0",
        "chunks": "Michelle Roman works at Gomez, Adams and Shepard as a Museum education officer. Michelle Roman was born on 1987-04-23 and lives at 03937 Andrew Mews Suite 199, Fisherton, SC 08281."
    }
]
Query took 47.20 seconds

Old
----------

Perform nilRAG...
Performance breakdown. Time to:            
 generate query embedding: 1.44 seconds             
 check clustering and (if existing) get closest centroid:                0.00 seconds            
 Number of clusters found: 0            
 encrypt query embedding: 0.00 seconds            
 ask nilDB to compute the differences: 63.55 seconds            
 group shares by Id: 0.00 seconds            
 transpose list: 2.03 seconds            
 decrypt: 13.61 seconds            
 get top chunks ids: 0.00 seconds            
 query top chunks: 0.46 seconds            
[
    {
        "_id": "7780ee4b-9b20-4bfd-afe2-d9be082a5e6f",
        "chunks": "Michelle Perez works at Gilmore, Vasquez and Francis as a Occupational psychologist. Michelle Perez was born on 2000-04-30 and lives at 2718 Burke Haven Suite 427, Tracieton, MD 74859."
    },
    {
        "_id": "fee6732f-c06e-4643-9f16-0a167592cfb0",
        "chunks": "Michelle Roman works at Gomez, Adams and Shepard as a Museum education officer. Michelle Roman was born on 1987-04-23 and lives at 03937 Andrew Mews Suite 199, Fisherton, SC 08281."
    }
]
Query took 81.28 seconds